LOC-63 Update likes counter

+++ app/Livewire/ProductFavorites.php
@@ -8,23 +8,34 @@ use Livewire\Component;
 
 class ProductFavorites extends Component
 {
-    public Product $product;
-    public function toggleFavorite(){
-          $user = auth()->user();
 
-          if($this->product->isFavoriteBy($user)){
-            Favorite::where('user_id' , $user->id)->where('product_id' , $this->product->id)->delete();
-          }else {
-            Favorite::create([
-                'user_id' => $user->id ,
-                'product_id' => $this->product->id,
-            ]);
-          }
-          $this->product->refresh();
-    }
+public ?Product $product = null;
 
-    public function render()
+    protected $listeners = ['load-product-favorites' => 'loadProduct'];
+    public function loadProduct($id)
     {
-        return view('livewire.product-favorites');
+        $this->product = Product::with('favorites')->find($id);
     }
+
+  public function toggleFavorite()
+  {
+    if (!$this->product) return;
+    $user = auth()->user();
+
+    if ($this->product->isFavoriteBy($user)) {
+      Favorite::where('user_id', $user->id)->where('product_id', $this->product->id)->delete();
+    } else {
+      Favorite::create([
+        'user_id' => $user->id,
+        'product_id' => $this->product->id,
+      ]);
+    }
+
+    $this->product->refresh();
+  }
+
+  public function render()
+  {
+    return view('livewire.product-favorites');
+  }
 }
